Refactor subcoin-utxo-snapshot
